move swapchain logic to separate files
