+++ src/include/graphics/vulkan/vk_types.h
@@ -25,6 +25,23 @@
         }                                                             \
     } while (0)
 
+struct DeletionQueue {
+    std::deque<std::function<void()>> deletors;
+
+    void push_function(std::function<void()>&& function) {
+        deletors.push_back(function);
+    }
+
+    void flush() {
+        // reverse iterate the deletion queue to execute all the functions
+        for (auto it = deletors.rbegin(); it != deletors.rend(); it++) {
+            (*it)();  // call functors
+        }
+
+        deletors.clear();
+    }
+};
+
 struct AllocatedImage {
     VkImage image;
     VkImageView imageView;
@@ -32,6 +49,20 @@ struct AllocatedImage {
     VkExtent3D imageExtent;
     VkFormat imageFormat;
 };
+
+struct VulkanContext {
+    VkExtent2D windowExtent{2560, 1440};
+    VkSurfaceKHR surface; // Vulkan window surface
+    VkPhysicalDevice chosenGPU; // GPU chosen as the default device
+    VkDevice device; // Vulkan device for commands
+
+    AllocatedImage drawImage;
+
+    DeletionQueue mainDeletionQueue;
+
+    bool resize_requested;
+};
+
 
 struct AllocatedBuffer {
     VkBuffer buffer;
